deps: update rustls 0.20.0 -> 0.21.0.
Patches rustls dependency from 0.20.0 to 0.21.0.

Also patches transitive deps on rustls-native-certs, reqwest,
hyper-rustls and tokio-rustls to use branches that also depend on rustls
0.21.0.

Most notable for this codebase, the
`rustls::Error::InvalidCertificateData` has been removed and replaced
by `InvalidCertificate` and a number of `CertificateError` sub-variants.

Now that the upstream `InvalidCertificate` error offers a way to specify
what went wrong with more granularity we're able to trim most of the
`error_messages` consts, choosing instead to return the more specific
upstream error types. For the case where invalid extensions are detected
the non-webpki platforms retain a custom error message const.

In the case where we want to return an `InvalidCertificate` error with
some platform specific error message we use the
`InvalidCertificate(CertificateError::Other)` variant with the error
message, repurposing the `invalid_certificate` helper for constructing
an `Arc` over a `Box` with the error message.

We also have to take some special care when asserting the equality of
errors, handling the case where looking at a `CertificateError::Other`
specially, since it can box a dyn error that can't be compared directly.