deps: update rustls 0.20.0 -> 0.21.0.
Patches rustls dependency from 0.20.0 to 0.21.0.

Also patches transitive deps on rustls-native-certs, reqwest,
hyper-rustls and tokio-rustls to use branches that also depend on rustls
0.21.0.

Two upstream changes have particular effect on this codebase:

* `rustls::Error` no longer implements `PartialEq` so `assert_eq!(...)`
  usage for comparing errors changes to `assert!(matches!(...))`
* `rustls::Error::InvalidCertificateData` has been removed and replaced
  by `InvalidCertificate` and a number of `CertificateError`
  subvariants.

Now that the upstream `InvalidCertificate` error offers a way to specify
what went wrong with more granularity we're able to remove the
`error_messages` consts, choosing instead to return the more specific
upstream error types.

In the case where we want to return an `InvalidCertificate` error with
some platform specific error message we use the
`InvalidCertificate(CertificateError::Other)` variant with the error
message, repurposing the `invalid_certificate` helper for constructing
an `Arc` over a `Box` with the error message.